Web Crawler ( Coursera Assignment Project)
The aim of this project is to crawl data from a web page provided by the professor and make some calculation on the data acquired. This project shold crwal the number in the column "comment" on http://py4e-data.dr-chuck.net/comments_1209104.html. Then it should calculate the sum of these numbers and print it.